requirements.txt
django
djangorestframework
Pillow
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
environs
coverage
openai
pandas
requests
bs4